Extend copy plugin to copy out files from pyodide
Updated webpack config to fix module output and handle node target.
Using the nodeExternals package in order to automatically externalize
all imports into the plugin so we do not need to maintain a list.

The goal is to use webpack to build the plugin for webpack. We do
not want to use it to bundle in external dependencies.