Write a plugin to copy necessary files for pyodide
Handle writing files to webpack output destination where they can
be loaded by emscripten based on the build configuration. Files
are copied to <dest>/pyodide/*.

A couple of options exist for the plugin.

autoCdnIndexUrl sets the cdn url for micropip and dependencies.
globalLoadPyodide sets whether or not loadPyodide is exposed globally