Change autoCdnUrl to micropipCdnUrl
Change due to pre-release and open issue #4 noting that users may
want to change the url. Having the value be a boolean would mean
setting the url would require a third option.

Opted to set the value as a string instead and default to the
release url. If users want to change it they can set it to whatever
url they would like.