Introduce a loader to pull in pyodide source
Reduces the amount of other configuration that may need to go
in like stubbing node core, ignoring certain node modules, or
writing your own loader.

Use acorn to find and replace global pyodide and add in named
exports so we have a consistent interface between common and esm.
Regular expression isn't as resiliant to change as AST.

CLean up configuration so it's easier to understand the different
build outputs for debugging.

Add examples for users.