Update webpack configs for build targets
Main change adds the after build into a new script to be
reused in the other configs. After build is used to copy the build
into the example folders which are useful for testing. You cannot
npm link the build because the links will resolve pyodide differently
and a check for specific types will fail when it loads the same file
twice from different location (they receive different memory addresses).