Add esbuild to required dependencies
Using esbuild to transpile the pyodide module (.mjs) to common
js format. There may be a better way to do this but the issue is:

When you build a module project (ie your project using pyodide is a
module) with webpack then it loads the pyodide.mjs.

We do not want webpack to process the pyodide.m?js file at all because
they have been output by emscripten and are already compatible with
the browser and node and contain pollyfills. The easiest way to handle
skipping a pollyfill is to load the file as a string and eval it
into existence. Webpack will skip processing this and treat it
as a string.

You cannot eval a module. eval doesn't recognize import/export. I
convert the build to commonjs, eval it, then export the loadPyodide
function back for esm compatability.