Smarter lookup of pyodide package location
We need to know where to copy files from so resolving the pyodide
package location is necessary. The old way assumed there was a
node_modules folder in your cwd and then listed all folders
in there looking for 1 named pyodide.

Now using require.resolve to let node find the package for us then
walk up to find the root of the package. Require.resolve returns
the main / entrypoint of the package which can be nested. We need
the folder that contains package.json.

There is a catch to stop traversing any lower than the node_modules
folder so it doesn't try to walk into protected directories.