Fixes #25 Ensure version exists in exports
Unfortunately I'm manually maintaining the exports for now.

loadPyodide and version are the only two things I think pyodide has
ever exports. In the future this could be automated better with
a more robust parser. I think I said that in several commits :). Let's
see how long this holds up and/or when I get some time to handle this
better. Ideally we'd live in a world that doesn't have commonjs
and esm anymore.